\@@_declare_shapes_smcaps:nn setup a `slsc -> itsc` substitution
If a slanted font is not provided, fontspec setup automatic `sl -> it`
substitutions in `\@@_declare_shape_slanted:nn`. But if a small caps
font was provided, an `itsc` shape is defined but there is no
corresponding `slsc -> itsc` substitution.

Fix this by adapting the code of `\@@_declare_shape_slanted:nn` into
`\@@_declare_shapes_smcaps:nn`.